requirements.txt
streamlit
gdown
tensorflow
opencv-python
numpy
Pillow

